Move the endpoints out
