+++ lsproxy/src/handlers/mod.rs
@@ -1,0 +1,6 @@
+mod workspace_files;
+mod references;
+mod definition;
+mod file_symbols;
+
+pub use self::{workspace_files::*, references::*, definition::*, file_symbols::*};
